requirements.txt
altgraph==0.17.3
certifi==2022.9.24
charset-normalizer==2.1.1
future==0.18.2
idna==3.4
iso8601==1.1.0
Nuitka==1.1.2
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyserial==3.5
PySide6==6.2.4
PyYAML==6.0
requests==2.28.1
shiboken6==6.2.4
urllib3==1.26.12
